requirements.txt
PyGithub>=1.55
aiohttp>=3.8.0
pyyaml>=5.1